M6-L6 Lesson and Practice Completed
During the practice, it was difficult to know what to use to add the candy as a child of the balloon. Using primarily the solution, I took note of what errors I committed and made sure that I learned about how to deal with something like this next time I encounter it.