M6 L8-L9 Reading and Experiments Completed
Putting some extra features on the dungeon scene. Added a bird that hops around, torches that turn on and turn off when clicked on, and collectibles that disappear when clicked on.

It was tough, reinforcing everything we learned, but a good challenge always gets us to think outside the box and try harder!